refactor : refactor the project structure
