+++ src/main/java/dev/example/xpenstracker/controller/BudgetController.java
@@ -1,0 +1,4 @@
+package dev.example.xpenstracker.controller;
+
+public class BudgetController {
+}
