--- src/main/java/dev/example/xpenstracker/model/ExpenseCategory.java
@@ -1,28 +1,0 @@
-package dev.example.xpenstracker.model;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "EXPENSE_CATEGORY")
-public class ExpenseCategory {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "CATEGORY_ID")
-    private Long categoryId;
-
-    @NotBlank(message = "CategoryName should not be empty")
-    @Column(name = "CATEGORY_NAME", nullable = false)
-
-    private String categoryName;
-
-
-}
