--- src/main/java/dev/example/xpenstracker/repository/ExpenseCategoryRepository.java
@@ -1,4 +1,0 @@
-package dev.example.xpenstracker.repository;
-
-public interface ExpenseCategoryRepository {
-}
