--- src/main/java/dev/example/xpenstracker/restController/BudgetController.java
@@ -1,4 +1,0 @@
-package dev.example.xpenstracker.restController;
-
-public class BudgetController {
-}
