--- src/main/java/dev/example/xpenstracker/service/ExpenseCategoryService.java
@@ -1,4 +1,0 @@
-package dev.example.xpenstracker.service;
-
-public class ExpenseCategoryService {
-}
